requirements.txt
streamlit
pypdf
langchain
faiss_cpu
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
flask